examples/requirements.txt
Django~=4.2.3
django-configurations~=2.4
python-dotenv~=1.0
psycopg2~=2.9
factory-boy~=3.3